+++ app/Http/Controllers/UserController.php
@@ -4,14 +4,15 @@ declare(strict_types=1);
 
 namespace App\Http\Controllers;
 
-use App\Http\Requests\ProfileUpdateRequest;
-use App\Jobs\DownloadUserAvatar;
+use App\Http\Requests\UserUpdateRequest;
+use App\Jobs\IncrementViews;
+use App\Jobs\UpdateUserAvatar;
 use App\Models\User;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use Illuminate\View\View;
 
-final readonly class ProfileController
+final readonly class UserController
 {
     /**
      * Display the user's profile form.
@@ -28,6 +29,8 @@ final readonly class ProfileController
      */
     public function show(User $user): View
     {
+        IncrementViews::dispatchUsingSession($user);
+
         return view('profile.show', [
             'user' => $user,
         ]);
@@ -36,7 +39,7 @@ final readonly class ProfileController
     /**
      * Update the user's profile information.
      */
-    public function update(ProfileUpdateRequest $request): RedirectResponse
+    public function update(UserUpdateRequest $request): RedirectResponse
     {
         $user = type($request->user())->as(User::class);
 
@@ -48,8 +51,13 @@ final readonly class ProfileController
 
         $user->save();
 
-        dispatch(new DownloadUserAvatar($user));
+        if ($user->wasChanged('email')) {
+            $user->sendEmailVerificationNotification();
 
+            if (! $user->is_uploaded_avatar) {
+                UpdateUserAvatar::dispatch($user);
+            }
+        }
         session()->flash('flash-message', 'Profile updated.');
 
         return to_route('profile.edit');