+++ app/Livewire/Home/QuestionsForYou.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace App\Livewire\Home;
 
+use App\Jobs\IncrementViews;
 use App\Livewire\Concerns\HasLoadMore;
 use App\Models\User;
 use App\Queries\Feeds\QuestionsForYouFeed;
@@ -21,10 +22,12 @@ final class QuestionsForYou extends Component
     {
         $user = type(auth()->user())->as(User::class);
 
-        $feed = new QuestionsForYouFeed($user);
+        $questions = (new QuestionsForYouFeed($user))->builder()->simplePaginate($this->perPage);
+
+        IncrementViews::dispatchUsingSession($questions->getCollection());
 
         return view('livewire.home.questions-for-you', [
-            'forYouQuestions' => $feed->builder()->simplePaginate($this->perPage),
+            'forYouQuestions' => $questions,
         ]);
     }
 }