+++ app/Livewire/Home/TrendingQuestions.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace App\Livewire\Home;
 
+use App\Jobs\IncrementViews;
 use App\Queries\Feeds\TrendingQuestionsFeed;
 use Illuminate\Http\Request;
 use Illuminate\View\View;
@@ -16,10 +17,12 @@ final class TrendingQuestions extends Component
      */
     public function render(Request $request): View
     {
-        $feed = new TrendingQuestionsFeed();
+        $questions = (new TrendingQuestionsFeed())->builder()->get();
+
+        IncrementViews::dispatchUsingSession($questions);
 
         return view('livewire.home.trending-questions', [
-            'trendingQuestions' => $feed->builder()->get(),
+            'trendingQuestions' => $questions,
         ]);
     }
 }