+++ app/Livewire/Links/Index.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace App\Livewire\Links;
 
-use App\Jobs\DownloadUserAvatar;
+use App\Jobs\UpdateUserAvatar;
 use App\Models\Link;
 use App\Models\User;
 use Illuminate\Auth\Access\AuthorizationException;
@@ -43,24 +43,6 @@ final class Index extends Component
     }
 
     /**
-     * Reset the user's avatar.
-     */
-    public function resetAvatar(): void
-    {
-        $user = type(auth()->user())->as(User::class);
-
-        if (! $this->canResetAvatar($user)) {
-            $this->dispatch('notification.created', message: 'You have to wait 24 hours before resetting the avatar again.');
-
-            return;
-        }
-
-        dispatch_sync(new DownloadUserAvatar($user));
-
-        $this->dispatch('notification.created', message: 'Avatar reset.');
-    }
-
-    /**
      * Store the new order of the links.
      *
      * @param  array<int, string>  $sort
@@ -93,9 +75,11 @@ final class Index extends Component
 
         $this->authorize('delete', $link);
 
-        dispatch(new DownloadUserAvatar($user));
-
         $link->delete();
+
+        if (! $user->is_uploaded_avatar) {
+            UpdateUserAvatar::dispatch($user);
+        }
 
         $this->dispatch('notification.created', message: 'Link deleted.');
     }
@@ -120,7 +104,6 @@ final class Index extends Component
 
         return view('livewire.links.index', [
             'user' => $user,
-            'canResetAvatar' => $this->canResetAvatar($user),
             'questionsReceivedCount' => $user->questionsReceived()
                 ->where('is_reported', false)
                 ->where('is_ignored', false)
@@ -133,16 +116,5 @@ final class Index extends Component
                 return $index;
             })->values(),
         ]);
-    }
-
-    /**
-     * Determine if the user can reset the avatar.
-     */
-    private function canResetAvatar(User $user): bool
-    {
-        return auth()->id() === $this->userId && (
-            $user->avatar_updated_at === null
-            || $user->avatar_updated_at->diffInHours(now()) > 24
-        );
     }
 }