+++ resources/views/livewire/links/index.blade.php
@@ -1,4 +1,8 @@
-<div>
+<div @if (auth()->user()?->is($user)) x-data="{
+    showSettingsForm: {{ $errors->settings->isEmpty() ? 'false' : 'true' }},
+    gradient: '{{ $user->gradient }}',
+    link_shape: '{{ $user->link_shape }}',
+}" @endif>
     <div class="relative bg-gradient-to-r p-5 text-center text-white">
         <div class="absolute left-0 top-6 flex">
             <button
@@ -40,20 +44,18 @@
 
         <div class="relative mx-auto h-24 w-24">
             <img
-                src="{{ $user->avatar ? url($user->avatar) : $user->avatar_url }}"
+                src="{{ $user->avatar_url }}"
                 alt="{{ $user->username }}"
                 class="{{ $user->is_company_verified ? 'rounded-md' : 'rounded-full' }} mx-auto mb-3 size-24"
             />
-            @if ($canResetAvatar)
+            @if (auth()->user()?->is($user))
                 <button
                     class="absolute right-0 top-0 rounded bg-slate-900 text-slate-300 transition duration-150 ease-in-out hover:bg-slate-800 hover:text-white"
-                    wire:click="resetAvatar"
-                    wire:loading.class="animate-spin"
-                    wire:target="resetAvatar"
-                    wire:loading.attr="disabled"
-                    title="Reset Avatar"
+                    href="{{ route('profile.edit') }}"
+                    wire:navigate
+                    title="Upload Avatar"
                 >
-                    <x-icons.reset class="size-5" />
+                    <x-icons.camera class="size-5" />
                 </button>
             @endif
         </div>
@@ -62,28 +64,57 @@
             <h2 class="text-2xl font-bold">{{ $user->name }}</h2>
 
             @if ($user->is_verified && $user->is_company_verified)
-                <x-icons.verified-company :color="$user->right_color" class="ml-1.5 size-6" />
+                <x-icons.verified-company
+                    :color="$user->right_color"
+                    class="ml-1.5 size-6"
+                />
             @elseif ($user->is_verified)
-                <x-icons.verified :color="$user->right_color" class="ml-1.5 size-6" />
+                <x-icons.verified
+                    :color="$user->right_color"
+                    class="ml-1.5 size-6"
+                />
             @endif
         </div>
 
-        <a class="text-slate-400" href="{{ route('profile.show', ['username' => $user->username]) }}" wire:navigate>
+        <a
+            class="text-slate-400"
+            href="{{ route('profile.show', ['username' => $user->username]) }}"
+            wire:navigate
+        >
             <p class="text-sm">{{ '@'.$user->username }}</p>
         </a>
 
         @if ($user->bio)
             <p class="text-sm">{{ $user->bio }}</p>
         @elseif (auth()->user()?->is($user))
-            <a href="{{ route('profile.edit') }}" class="text-sm text-slate-500 hover:underline" wire:navigate>Tell people about yourself</a>
+            <a
+                href="{{ route('profile.edit') }}"
+                class="text-sm text-slate-500 hover:underline"
+                wire:navigate
+                >Tell people about yourself</a
+            >
         @endif
 
         <div class="mt-2 text-sm">
             <p class="text-slate-400">
-                <span>
-                    {{ $questionsReceivedCount }}
+                <span
+                    class="cursor-help"
+                    title="{{ Number::format($questionsReceivedCount) }} {{ str('Answer')->plural($questionsReceivedCount) }}"
+                >
+                    {{ Number::abbreviate($questionsReceivedCount) }}
                     {{ str('Answer')->plural($questionsReceivedCount) }}
                 </span>
+
+                @if ($user->views > 0)
+                    <span class="mx-1">•</span>
+
+                    <span
+                        class="cursor-help"
+                        title="{{ Number::format($user->views) }} {{ str('view')->plural($user->views) }}"
+                    >
+                        {{ Number::abbreviate($user->views) }} {{ str('view')->plural($user->views) }}
+                    </span>
+                @endif
 
                 <span class="mx-1">•</span>
 
@@ -112,6 +143,7 @@
                     @foreach ($links as $link)
                         <li
                             class="{{ $user->link_shape }} {{ $user->gradient }} hover:darken-gradient group flex bg-gradient-to-r"
+                            :class="showSettingsForm && gradient + ' ' + link_shape"
                             x-sortable-item="{{ $link->id }}"
                             wire:key="link-{{ $link->id }}"
                         >
@@ -122,7 +154,10 @@
                                 <x-icons.sortable-handle class="size-6 opacity-100 group-hover:opacity-100 sm:opacity-0" />
                             </div>
 
-                            <x-links.list-item :$user :$link />
+                            <x-links.list-item
+                                :$user
+                                :$link
+                            />
 
                             <div class="flex items-center justify-center">
                                 <div
@@ -155,7 +190,10 @@
                             class="{{ $user->link_shape }} {{ $user->gradient }} hover:darken-gradient flex bg-gradient-to-r"
                             wire:click="click({{ $link->id }})"
                         >
-                            <x-links.list-item :$user :$link />
+                            <x-links.list-item
+                                :$user
+                                :$link
+                            />
                         </div>
                     @endforeach
                 </div>
@@ -167,7 +205,6 @@
         <div
             x-data="{
                 showLinksForm: {{ $errors->links->isEmpty() ? 'false' : 'true' }},
-                showSettingsForm: {{ $errors->settings->isEmpty() ? 'false' : 'true' }},
             }"
             class="py-4"
         >
@@ -175,14 +212,16 @@
                 <div class="flex gap-2">
                     <button
                         x-on:click="showLinksForm = ! showLinksForm ; showSettingsForm = false"
-                        class="bg-{{ $user->left_color }} {{ $user->link_shape }} hover:darken-gradient flex w-full basis-4/5 items-center justify-center px-4 py-2 text-sm font-bold text-white transition duration-300 ease-in-out"
+                        class="{{ $user->gradient }} {{ $user->link_shape }} hover:darken-gradient flex w-full basis-4/5 items-center justify-center bg-gradient-to-r px-4 py-2 text-sm font-bold text-white transition duration-300 ease-in-out"
+                        :class="showSettingsForm && gradient + ' ' + link_shape"
                     >
                         <x-icons.plus class="mr-1.5 size-5" />
                         Add New Link
                     </button>
                     <button
                         x-on:click="showSettingsForm = ! showSettingsForm ; showLinksForm = false"
-                        class="{{ $user->gradient }} hover:darken-gradient {{ $user->link_shape }} flex w-full basis-1/5 items-center justify-center bg-gradient-to-r px-4 py-2 font-bold text-white transition duration-300 ease-in-out"
+                        class="bg-{{ $user->right_color }} hover:darken-gradient {{ $user->link_shape }} flex w-full basis-1/5 items-center justify-center px-4 py-2 font-bold text-white transition duration-300 ease-in-out"
+                        :class="showSettingsForm && 'bg-' + gradient.split(' ')[1].replace('to-', '') + ' ' + link_shape"
                     >
                         <x-icons.cog class="size-6" />
                     </button>