+++ resources/views/livewire/questions/show.blade.php
@@ -17,7 +17,7 @@
                 >
                     <figure class="{{ $question->from->is_company_verified ? 'rounded-md' : 'rounded-full' }} h-10 w-10 flex-shrink-0 bg-slate-800 transition-opacity group-hover:opacity-90">
                         <img
-                            src="{{ $question->from->avatar ? url($question->from->avatar) : $question->from->avatar_url }}"
+                            src="{{ $question->from->avatar_url }}"
                             alt="{{ $question->from->username }}"
                             class="{{ $question->from->is_company_verified ? 'rounded-md' : 'rounded-full' }} h-10 w-10"
                         />
@@ -30,9 +30,15 @@
                             </p>
 
                             @if ($question->from->is_verified && $question->from->is_company_verified)
-                                <x-icons.verified-company :color="$question->from->right_color" class="ml-1 mt-0.5 h-3.5 w-3.5" />
+                                <x-icons.verified-company
+                                    :color="$question->from->right_color"
+                                    class="ml-1 mt-0.5 h-3.5 w-3.5"
+                                />
                             @elseif ($question->from->is_verified)
-                                <x-icons.verified :color="$question->from->right_color" class="ml-1 mt-0.5 h-3.5 w-3.5" />
+                                <x-icons.verified
+                                    :color="$question->from->right_color"
+                                    class="ml-1 mt-0.5 h-3.5 w-3.5"
+                                />
                             @endif
                         </div>
 
@@ -58,10 +64,14 @@
     @if ($question->answer)
         <div class="answer mt-3 rounded-2xl bg-slate-900 p-4">
             <div class="flex justify-between">
-                <a href="{{ route('profile.show', ['username' => $question->to->username]) }}" class="group flex items-center gap-3" wire:navigate>
+                <a
+                    href="{{ route('profile.show', ['username' => $question->to->username]) }}"
+                    class="group flex items-center gap-3"
+                    wire:navigate
+                >
                     <figure class="{{ $question->to->is_company_verified ? 'rounded-md' : 'rounded-full' }} h-10 w-10 flex-shrink-0 bg-slate-800 transition-opacity group-hover:opacity-90">
                         <img
-                            src="{{ $question->to->avatar ? url($question->to->avatar) : $question->to->avatar_url }}"
+                            src="{{ $question->to->avatar_url }}"
                             alt="{{ $question->to->username }}"
                             class="{{ $question->to->is_company_verified ? 'rounded-md' : 'rounded-full' }} h-10 w-10"
                         />
@@ -73,9 +83,15 @@
                             </p>
 
                             @if ($question->to->is_verified && $question->to->is_company_verified)
-                                <x-icons.verified-company :color="$question->to->right_color" class="ml-1 mt-0.5 h-3.5 w-3.5" />
+                                <x-icons.verified-company
+                                    :color="$question->to->right_color"
+                                    class="ml-1 mt-0.5 h-3.5 w-3.5"
+                                />
                             @elseif ($question->to->is_verified)
-                                <x-icons.verified :color="$question->to->right_color" class="ml-1 mt-0.5 h-3.5 w-3.5" />
+                                <x-icons.verified
+                                    :color="$question->to->right_color"
+                                    class="ml-1 mt-0.5 h-3.5 w-3.5"
+                                />
                             @endif
                         </div>
 
@@ -85,7 +101,10 @@
                     </div>
                 </a>
                 @if (auth()->check() && auth()->user()->can('update', $question))
-                    <x-dropdown align="right" width="48">
+                    <x-dropdown
+                        align="right"
+                        width="48"
+                    >
                         <x-slot name="trigger">
                             <button class="inline-flex items-center rounded-md border border-transparent py-1 text-sm text-slate-400 transition duration-150 ease-in-out hover:text-slate-50 focus:outline-none">
                                 <x-icons.ellipsis-horizontal class="h-6 w-6" />
@@ -94,12 +113,18 @@
 
                         <x-slot name="content">
                             @if (! $question->pinned && auth()->user()->can('pin', $question))
-                                <x-dropdown-button wire:click="pin" class="flex items-center gap-1.5">
+                                <x-dropdown-button
+                                    wire:click="pin"
+                                    class="flex items-center gap-1.5"
+                                >
                                     <x-icons.pin class="h-4 w-4 text-slate-50" />
                                     <span>Pin</span>
                                 </x-dropdown-button>
                             @elseif ($question->pinned)
-                                <x-dropdown-button wire:click="unpin" class="flex items-center gap-1.5">
+                                <x-dropdown-button
+                                    wire:click="unpin"
+                                    class="flex items-center gap-1.5"
+                                >
                                     <x-icons.pin class="h-4 w-4" />
                                     <span>Unpin</span>
                                 </x-dropdown-button>
@@ -143,9 +168,24 @@
 
                         @php($likesCount = $question->likes_count)
                         @if ($likesCount)
-                            <p class="cursor-click ml-1" title="{{ Number::format($likesCount) }} {{ str('like')->plural($likesCount) }}">{{ Number::abbreviate($likesCount) }} {{ str('like')->plural($likesCount) }}</p>
+                            <p
+                                class="cursor-click ml-1"
+                                title="{{ Number::format($likesCount) }} {{ str('like')->plural($likesCount) }}"
+                            >
+                                {{ Number::abbreviate($likesCount) }} {{ str('like')->plural($likesCount) }}
+                            </p>
                         @endif
                     </button>
+                    @if ($question->views > 0)
+                        <span class="mx-1">•</span>
+                        <x-icons.chart class="h-4 w-4" />
+                        <p
+                            class="ml-1 cursor-help"
+                            title="{{ Number::format($question->views) }} {{ str('view')->plural($question->views) }}"
+                        >
+                            {{ Number::abbreviate($question->views) }} {{ str('view')->plural($question->views) }}
+                        </p>
+                    @endif
                 </div>
                 <div class="flex items-center text-slate-500">
                     <time
@@ -201,6 +241,9 @@
             </div>
         </div>
     @elseif (auth()->user()?->is($user))
-        <livewire:questions.edit :questionId="$question->id" :key="$question->id" />
+        <livewire:questions.edit
+            :questionId="$question->id"
+            :key="$question->id"
+        />
     @endif
 </article>