+++ routes/web.php
@@ -4,12 +4,13 @@ declare(strict_types=1);
 
 use App\Http\Controllers\ChangelogController;
 use App\Http\Controllers\NotificationController;
-use App\Http\Controllers\Profile\Connect\GitHubController;
-use App\Http\Controllers\Profile\TimezoneController;
-use App\Http\Controllers\Profile\VerifiedController;
-use App\Http\Controllers\ProfileController;
 use App\Http\Controllers\QrCodeController;
 use App\Http\Controllers\QuestionController;
+use App\Http\Controllers\UserAvatarController;
+use App\Http\Controllers\UserController;
+use App\Http\Controllers\UserGitHubUsernameController;
+use App\Http\Controllers\UserIsVerifiedController;
+use App\Http\Controllers\UserTimezoneController;
 use App\Http\Middleware\EnsureVerifiedEmailsForSignInUsers;
 use Illuminate\Support\Facades\Route;
 
@@ -28,10 +29,10 @@ Route::view('/brand/resources', 'brand.resources')->name('brand.resources');
 Route::redirect('/sponsors', 'https://github.com/sponsors/nunomaduro/')->name('sponsors');
 
 Route::get('/changelog', [ChangelogController::class, 'show'])->name('changelog');
-Route::post('/profile/timezone', [TimezoneController::class, 'update'])->name('profile.timezone.update');
+Route::post('/profile/timezone', [UserTimezoneController::class, 'update'])->name('profile.timezone.update');
 
 Route::prefix('/@{username}')->group(function () {
-    Route::get('/', [ProfileController::class, 'show'])
+    Route::get('/', [UserController::class, 'show'])
         ->name('profile.show')
         ->middleware(EnsureVerifiedEmailsForSignInUsers::class);
 
@@ -44,29 +45,34 @@ Route::group(['middleware' => ['auth', 'verified']], function () {
     Route::get('notifications', [NotificationController::class, 'index'])->name('notifications.index');
     Route::get('notifications/{notification}', [NotificationController::class, 'show'])
         ->name('notifications.show');
+
+    Route::patch('/profile/avatar', [UserAvatarController::class, 'update'])
+        ->name('profile.avatar.update');
+    Route::delete('/profile/avatar', [UserAvatarController::class, 'delete'])
+        ->name('profile.avatar.delete');
 });
 
 Route::middleware('auth')->group(function () {
     Route::get('/qr-code', QrCodeController::class)->name('qr-code.image');
 
-    Route::get('/profile', [ProfileController::class, 'edit'])->name('profile.edit');
-    Route::patch('/profile', [ProfileController::class, 'update'])->name('profile.update');
-    Route::delete('/profile', [ProfileController::class, 'destroy'])->name('profile.destroy');
+    Route::get('/profile', [UserController::class, 'edit'])->name('profile.edit');
+    Route::patch('/profile', [UserController::class, 'update'])->name('profile.update');
+    Route::delete('/profile', [UserController::class, 'destroy'])->name('profile.destroy');
 
     Route::prefix('/profile/connect/github')->group(function () {
-        Route::get('/', [GitHubController::class, 'index'])
+        Route::get('/', [UserGitHubUsernameController::class, 'index'])
             ->name('profile.connect.github');
 
         Route::get('/update', [
-            GitHubController::class, 'update',
+            UserGitHubUsernameController::class, 'update',
         ])->name('profile.connect.github.update');
 
         Route::delete('/', [
-            GitHubController::class, 'destroy',
+            UserGitHubUsernameController::class, 'destroy',
         ])->name('profile.connect.github.destroy');
     });
 
-    Route::post('/profile/verified', [VerifiedController::class, 'update'])
+    Route::post('/profile/verified', [UserIsVerifiedController::class, 'update'])
         ->name('profile.verified.update');
 });