+++ tests/Arch/GlobalTest.php
@@ -3,7 +3,7 @@
 declare(strict_types=1);
 
 arch('globals')
-    ->expect(['dd', 'dump', 'ray', 'die', 'var_dump', 'sleep'])
+    ->expect(['dd', 'dump', 'ray', 'die', 'var_dump', 'sleep', 'dispatch', 'dispatch_sync'])
     ->not->toBeUsed();
 
 arch('http helpers')
@@ -11,4 +11,5 @@ arch('http helpers')
     ->toOnlyBeUsedIn([
         'App\Http',
         'App\Livewire',
+        'App\Jobs\IncrementViews',
     ]);