+++ tests/Http/Home/ForYouTest.php
@@ -2,6 +2,7 @@
 
 declare(strict_types=1);
 
+use App\Jobs\IncrementViews;
 use App\Livewire\Home\QuestionsForYou;
 use App\Models\User;
 
@@ -20,4 +21,13 @@ it('guest can see the "for you" view', function () {
 
     $response->assertOk()
         ->assertSee('Log in or sign up to access personalized content');
+});
+
+it('does increment views', function () {
+    Queue::fake(IncrementViews::class);
+
+    $this->actingAs(User::factory()->create());
+    $this->get(route('home.for_you'));
+
+    Queue::assertPushed(IncrementViews::class);
 });