+++ tests/Http/Register/CreateTest.php
@@ -25,6 +25,7 @@ test('new users can register', function () {
         'email' => 'test@example.com',
         'password' => 'm@9v_.*.XCN',
         'password_confirmation' => 'm@9v_.*.XCN',
+        'terms' => true,
         'g-recaptcha-response' => 'valid',
     ]);
 
@@ -52,7 +53,7 @@ test('required fields', function (string $field) {
         ->assertSessionHasErrors([
             $field => 'The '.$field.' field is required.',
         ]);
-})->with(['name', 'username', 'email', 'password']);
+})->with(['name', 'username', 'email', 'password', 'terms']);
 
 test('email must be valid', function () {
     $response = $this->from('/register')->post('/register', [
@@ -78,6 +79,20 @@ test('password must be confirmed', function () {
 
     $response->assertRedirect('/register')
         ->assertSessionHasErrors(['password' => 'The password field confirmation does not match.']);
+});
+
+test('users must be at least 18 years old', function () {
+    $response = $this->from('/register')->post('/register', [
+        'name' => 'Test User',
+        'username' => 'testuser',
+        'email' => 'test@example.com',
+        'password' => 'password',
+        'password_confirmation' => 'not-password',
+        'terms' => false,
+    ]);
+
+    $response->assertRedirect('/register')
+        ->assertSessionHasErrors(['terms' => 'The terms field must be accepted.']);
 });
 
 test('username must be unique', function () {
@@ -255,6 +270,7 @@ test("user's name can contain blank characters", function (string $given, string
         'email' => 'test@laravel.com',
         'password' => 'password',
         'password_confirmation' => 'password',
+        'terms' => true,
     ]);
 
     $user = User::where('email', 'test@laravel.com')->first();
@@ -273,6 +289,7 @@ test('anonymously preference is set to true by default', function () {
         'email' => 'test@example.com',
         'password' => 'password',
         'password_confirmation' => 'password',
+        'terms' => true,
     ]);
 
     expect(User::first()->prefers_anonymous_questions)->toBeTrue();