+++ CHANGELOG.md
@@ -1,3 +1,31 @@
+## Version 1.12.0
+> 24 April, 2024
+
+- Improves avatar image fetching from social media.
+
+
+## Version 1.11.0
+> 23 April, 2024
+
+- Improves Firewall system to mitigate the growing volume of profile and question views by bots.
+
+## Version 1.10.0
+> 22 April, 2024
+
+- Adds "Back To Top" button to the user's profile page and feeds.
+- Adds "Accept Terms" checkbox to the user's registration form.
+- Makes mastodon links verifiable.
+
+## Version 1.9.0
+> 19 April, 2024
+
+- Adds possibility to upload a custom avatar image to the user's profile.
+
+## Version 1.8.0
+> 17 April, 2024
+
+- Added views count to the user's profile page and questions / answers blocks.
+
 ## Version 1.7.0
 > 13 April, 2024